refactoring: change name
