+++ app/controllers/PostProcess.scala
@@ -2,7 +2,7 @@ package controllers
 
 import com.daumkakao.s2graph.core._
 import com.daumkakao.s2graph.core.mysqls._
-import com.daumkakao.s2graph.core.types2.{InnerVal, InnerValLike}
+import com.daumkakao.s2graph.core.types.{InnerVal, InnerValLike}
 import com.daumkakao.s2graph.logger
 import play.api.libs.json.{JsObject, JsValue, Json}
