+++ s2core/src/main/scala/com/daumkakao/s2graph/core/parsers/WhereParser.scala
@@ -2,7 +2,7 @@ package com.daumkakao.s2graph.core.parsers
 
 import com.daumkakao.s2graph.core._
 import com.daumkakao.s2graph.core.mysqls._
-import com.daumkakao.s2graph.core.types2.InnerValLike
+import com.daumkakao.s2graph.core.types.InnerValLike
 
 import scala.util.Try
 import scala.util.parsing.combinator.JavaTokenParsers
