+++ s2core/src/main/scala/com/daumkakao/s2graph/core/types/EdgeValueLike.scala
@@ -1,4 +1,4 @@
-package com.daumkakao.s2graph.core.types2
+package com.daumkakao.s2graph.core.types
 
 /**
  * Created by shon on 6/10/15.
