+++ s2core/src/main/scala/com/daumkakao/s2graph/core/types/v1/EdgeQualifierInverted.scala
@@ -1,10 +1,11 @@
-package com.daumkakao.s2graph.core.types2.v2
+package com.daumkakao.s2graph.core.types.v1
 
-import com.daumkakao.s2graph.core.types2._
+import com.daumkakao.s2graph.core.types._
 
 /**
  * Created by shon on 6/10/15.
  */
+
 object EdgeQualifierInverted extends HBaseDeserializable {
   import HBaseType._
   def fromBytes(bytes: Array[Byte],
