+++ s2core/src/main/scala/com/daumkakao/s2graph/core/types/v1/EdgeValue.scala
@@ -1,17 +1,17 @@
-package com.daumkakao.s2graph.core.types2.v2
+package com.daumkakao.s2graph.core.types.v1
 
-import com.daumkakao.s2graph.core.types2._
+import com.daumkakao.s2graph.core.types._
 
 /**
  * Created by shon on 6/10/15.
  */
 object EdgeValue extends HBaseDeserializable {
-  import HBaseType._
+import HBaseType._
   import HBaseDeserializable._
   def fromBytes(bytes: Array[Byte],
                 offset: Int,
                 len: Int,
-                version: String = VERSION2): (EdgeValue, Int) = {
+                version: String = VERSION1): (EdgeValue, Int) = {
     val (props, endAt) = bytesToKeyValues(bytes, offset, 0, version)
     (EdgeValue(props), endAt - offset)
   }
