+++ s2core/src/main/scala/com/daumkakao/s2graph/core/types/v1/EdgeValueInverted.scala
@@ -1,6 +1,6 @@
-package com.daumkakao.s2graph.core.types2.v1
+package com.daumkakao.s2graph.core.types.v1
 
-import com.daumkakao.s2graph.core.types2._
+import com.daumkakao.s2graph.core.types._
 import org.apache.hadoop.hbase.util.Bytes
 
 /**
