+++ s2core/src/main/scala/com/daumkakao/s2graph/core/types/v2/EdgeQualifier.scala
@@ -1,7 +1,7 @@
-package com.daumkakao.s2graph.core.types2.v2
+package com.daumkakao.s2graph.core.types.v2
 
 import com.daumkakao.s2graph.core.GraphUtil
-import com.daumkakao.s2graph.core.types2._
+import com.daumkakao.s2graph.core.types._
 import org.apache.hadoop.hbase.util.Bytes
 
 /**
