+++ s2core/src/test/scala/com/daumkakao/s2graph/core/parsers/WhereParserTest.scala
@@ -2,7 +2,7 @@ package com.daumkakao.s2graph.core.parsers
 
 import com.daumkakao.s2graph.core._
 //import com.daumkakao.s2graph.core.mysqls._
-import com.daumkakao.s2graph.core.types2._
+import com.daumkakao.s2graph.core.types._
 import org.scalatest.{Matchers, FunSuite}
 import play.api.libs.json.Json
